requirements.txt
discord.py>=2.3.2
firebase-admin>=6.4.0
